requirements.txt
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
ci-info==0.1.1
Click==7.0
cycler==0.10.0
decorator==4.4.2
dicom2nifti==2.2.6
dill==0.3.1.1
dominate==2.4.0
etelemetry==0.2.0
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
Logbook==1.5.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.2.0
multiprocess==0.70.9
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.2
nipype==1.4.2
numpy==1.18.2
packaging==20.3
pathlib==1.0.1
pluggy==0.13.1
prov==1.5.3
py==1.8.1
pydicom==1.4.2
pydot==1.4.1
pydotplus==2.0.2
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.12.0
rdflib==4.2.2
requests==2.23.0
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
traits==6.0.0
transliterate==1.10.2
urllib3==1.25.8
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.12.1
wrapt-timeout-decorator==1.3.1
WTForms==2.2.1
zipp==3.1.0
